Add files via upload
Want to avoid destroying the original  data file.  Solution to try is a second file (data.tmp).  This file will then contain the residual unsold list of items.  The data file itself will be unchanged.  I will then add the residual items to the spreadsheet output.  
I don't know if temporary files will work for this application.  This progam opens and closes the file frequently until the data is analized.